requirements.txt
fastapi==0.115.6
httpx==0.28.1
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
pandas==2.2.3
python-dotenv==1.0.1
rapidfuzz==3.10.1
tldextract==5.1.2
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic-ai-slim[google]==1.16.0
